requirements.txt
ansiwrap==0.8.4
asgiref==3.7.2
console-menu==0.8.0
Django==4.2.3
django-activity-stream==1.4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
textwrap3==0.9.2
typing_extensions==4.7.1
uuid==1.30

pip~=23.2
setuptools~=68.0.0
requests==2.31.0
